examples/cli-starter-template/requirements.txt
# CLI Starter Template Dependencies
#
# This template requires valid8r for input validation
# and PyYAML for configuration file parsing

valid8r>=1.19.0
PyYAML>=6.0.0
